services/users/requirements.txt
Flask==1.0.2
Flask_SQLAlchemy==2.3.2
psycopg2-binary==2.7.6.1
Flask_Testing==0.7.1
coverage==4.5.2
flake8==3.6.0
flask-debugtoolbar==0.10.1
flask-cors==3.0.7
flask-migrate==2.3.1
flask-bcrypt==0.7.1